Split theme preset data
